+++ database/migrations/2025_10_05_170705_add_almacen_id_to_inventario_movimientos_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('inventario_movimientos', function (Blueprint $table) {
-            $table->foreignId('almacen_id')->nullable()->after('producto_id')->constrained('almacenes')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('inventario_movimientos', 'almacen_id')) {
+            Schema::table('inventario_movimientos', function (Blueprint $table) {
+                $table->foreignId('almacen_id')->nullable()->after('producto_id')->constrained('almacenes')->onDelete('cascade');
+            });
+        }
     }
 
     /**
